+++ clients/js/src/confidential.ts
@@ -1,3 +1,2 @@
 export * from './confidentialTransferHelpers';
 export * from './confidentialTransferKeys';
-export * from './confidentialTransferProofTransactionPlanning';
